refactor #10 : 뉴스 기사 목록 조회
- Test를 위해 임의로 만든 Interest Entity 제거 후 도메인 Interest로 주입
