+++ src/test/java/org/project/monewping/domain/article/service/ArticlesServiceTest.java
@@ -34,7 +34,6 @@ import org.project.monewping.domain.article.exception.DuplicateArticleException;
 import org.project.monewping.domain.article.exception.InterestNotFoundException;
 import org.project.monewping.domain.article.mapper.ArticlesMapper;
 import org.project.monewping.domain.article.repository.ArticlesRepository;
-import org.project.monewping.domain.article.repository.InterestRepository;
 import org.project.monewping.global.dto.CursorPageResponse;
 import org.project.monewping.domain.interest.entity.Interest;
 import org.project.monewping.domain.interest.repository.InterestRepository;
